JSON part one (needs work)
I have created a big huge list of JSON but not all fields line up with
the database.  If we go the "all one big schema" route (which I favor),
this "order" may be the order we need?  There are plenty of notes in
application.js that need addressing at this time.  Thanks!  :)