vetting worksheet BEGINS
Holy Crap it Works.  Got routing to /vettingworksheet/:id to go to a new
page that I'll be continuing to work on - but this is a good start.  Did
a ton of copy paste that will be deleted.  Got a persistent vetting note
to work (but it doesn't look all that great...gotta figure out more
bootstrap CSS)